Fix four clang-tidy check categories
readability-math-missing-parentheses, bugprone-narrowing-conversions,
misc-unused-parameters and bugprone-too-small-loop-variable are now clean
in our own code (the remaining narrowing hits are in ext/cmark-gfm and the
WebView2 package).

Most of it is mechanical: explicit parentheses around * inside +/-, and
explicit casts at int/float boundaries. The substantive bits:

- kEofRec was an int holding 0xe98e0d0a, which doesn't fit; it is a u32
- the aboutHover class style is a ULONG_PTR, cast where it is passed back
- AnnotEditToolbar's font argument was threaded through three functions
  and never used, so the whole chain is gone

cmd/clang-tidy.ts was missing ext/msdes from includeDirs, so LitDoc.cpp
failed to parse on d3des.h and clang-tidy silently declined to apply any
fix to it.

prompt: fix clang-tidy readability-math-missing-parentheses, misc-unused-parameters, bugprone-narrowing-conversions, bugprone-too-small-loop-variable

+++ src/EngineImages.cpp
@@ -636,10 +636,10 @@ Pixmap* EngineImages::RenderPage(RenderPageArgs& args) {
         fz_irect subarea;
         fz_irect* subPtr = nullptr;
         if (!isFullPage && pageRect) {
-            subarea.x0 = pageRc.x;
-            subarea.y0 = pageRc.y;
-            subarea.x1 = pageRc.x + pageRc.dx;
-            subarea.y1 = pageRc.y + pageRc.dy;
+            subarea.x0 = (int)pageRc.x;
+            subarea.y0 = (int)pageRc.y;
+            subarea.x1 = (int)(pageRc.x + pageRc.dx);
+            subarea.y1 = (int)(pageRc.y + pageRc.dy);
             if (subarea.x0 < 0) {
                 subarea.x0 = 0;
             }
